refactor: share dual mode module
